dziala zapis

+++ Project/src/project/PropertiesHandler.java
@@ -22,20 +22,23 @@ public class PropertiesHandler {
     }
 
     public void saveProperty(Object key, String value) {
-        //try (FileOutputStream output = new FileOutputStream(fileName, true)) {
-        //    properties.setProperty(key, value);
-        //    properties.store(output, null);
-        try(BufferedWriter writer = new BufferedWriter(new FileWriter(fileName, true))) {
-        	String property = key + " = " + value;
-        	writer.write(property);
-        	writer.newLine();
-        	
-    	} catch (IOException e) {
-            e.printStackTrace();
-            //logowanie błędu do pliku
-            Logging logging = new Logging(PropertiesHandler.class);
-            logging.logError("An eror occured while saving property to file.", e);
+        if (!isValueExists(value)) {
+            try (BufferedWriter writer = new BufferedWriter(new FileWriter(fileName, true))) {
+                String property = key + " = " + value;
+                writer.write(property);
+                writer.newLine();
+                LOGGER.info("Property saved: " + property);
+            } catch (IOException e) {
+                e.printStackTrace();
+                LOGGER.error("An error occurred while saving property to file.", e);
+            }
+        } else {
+            LOGGER.info("Property already exists: " + value);
         }
+    }
+
+    private boolean isValueExists(String value) {
+        return properties.values().stream().anyMatch(existingValue -> existingValue.equals(value));
     }
 
     public String getProperty(String key) {
